+++ lib/puppet-check/data_parser.rb
@@ -188,10 +188,14 @@ class DataParser
     private_class_method :method
     warnings = []
 
-    data.each do |key, value|
-      # check for nil values in the data (nil keys are fine)
-      if (value.is_a?(Hash) && value.values.any?(&:nil?)) || value.nil?
-        warnings.push("Value(s) missing in key '#{key}'.")
+    # disregard nil/undef value data check if hiera >=5
+    # this constant is coming from puppet from puppet_parser from puppet_check from here
+    if Puppet::PUPPETVERSION.to_i < 5
+      data.each do |key, value|
+        # check for nil values in the data (nil keys are fine)
+        if (value.is_a?(Hash) && value.values.any?(&:nil?)) || value.nil?
+          warnings.push("Value(s) missing in key '#{key}'.")
+        end
       end
     end